- changed callback argument type from `uint8_t` to `int` to remove beginner's confusion with _obscure_ types - updated example, README and version number (1.3.2)

+++ examples/FTDebouncer_demo/FTDebouncer_demo.ino
@@ -40,11 +40,11 @@ void loop(){
 	usElapsed = micros() - us;
 	Serial.println(usElapsed);
 }
-void onPinActivated(uint8_t pinNr){
+void onPinActivated(int pinNr){
 	Serial.print("Pin activated: ");
 	Serial.println(pinNr);
 }
-void onPinDeactivated(uint8_t pinNr){
+void onPinDeactivated(int pinNr){
 	Serial.print("Pin deactivated: ");
 	Serial.println(pinNr);
 }
